# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/remainingnumber/reserveleave/TempReserveLeaveMngRepository.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/workrecord/erroralarm/EmployeeDailyPerErrorRepository.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/repository/remainingnumber/JpaTempReserveLeaveMngRepo.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/workrule/closure/Closure.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.infra/src/main/java/nts/uk/ctx/bs/employee/infra/repository/employee/mngdata/EmployeeDataMngInfoRepositoryImp.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pub/src/main/java/nts/uk/ctx/bs/employee/pub/employee/EmployeeInfoExport.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/add/AddSpLea13InfoCommandHandler.java
@@ -14,12 +14,13 @@ import nts.uk.shr.pereg.app.command.PeregAddCommandHandler;
 import nts.uk.shr.pereg.app.command.PeregAddCommandResult;
 
 @Stateless
-public class AddSpLea13InfoCommandHandler extends CommandHandlerWithResult<AddSpecialleave13informationCommand, PeregAddCommandResult>
-implements PeregAddCommandHandler<AddSpecialleave13informationCommand>{
+public class AddSpLea13InfoCommandHandler
+		extends CommandHandlerWithResult<AddSpecialleave13informationCommand, PeregAddCommandResult>
+		implements PeregAddCommandHandler<AddSpecialleave13informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler addSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00051";
@@ -32,12 +33,13 @@ implements PeregAddCommandHandler<AddSpecialleave13informationCommand>{
 
 	@Override
 	protected PeregAddCommandResult handle(CommandHandlerContext<AddSpecialleave13informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00051.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
+
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00051.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
 				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		return new PeregAddCommandResult(addSpLeaInfoCommandHandler.addHandler(domain));
 	}
